lsteamclient: Use standard fixed size integer types.
CW-Bug-Id: #22729

+++ lsteamclient/cppISteamNetworkingUtils_SteamNetworkingUtils001.h
@@ -9,7 +9,7 @@ extern void cppISteamNetworkingUtils_SteamNetworkingUtils001_GetLocalPingLocatio
 struct cppISteamNetworkingUtils_SteamNetworkingUtils001_EstimatePingTimeBetweenTwoLocations_params
 {
     void *linux_side;
-    int _ret;
+    int32_t _ret;
     const SteamNetworkPingLocation_t *location1;
     const SteamNetworkPingLocation_t *location2;
 };
@@ -18,7 +18,7 @@ extern void cppISteamNetworkingUtils_SteamNetworkingUtils001_EstimatePingTimeBet
 struct cppISteamNetworkingUtils_SteamNetworkingUtils001_EstimatePingTimeFromLocalHost_params
 {
     void *linux_side;
-    int _ret;
+    int32_t _ret;
     const SteamNetworkPingLocation_t *remoteLocation;
 };
 extern void cppISteamNetworkingUtils_SteamNetworkingUtils001_EstimatePingTimeFromLocalHost( struct cppISteamNetworkingUtils_SteamNetworkingUtils001_EstimatePingTimeFromLocalHost_params *params );
@@ -28,7 +28,7 @@ struct cppISteamNetworkingUtils_SteamNetworkingUtils001_ConvertPingLocationToStr
     void *linux_side;
     const SteamNetworkPingLocation_t *location;
     char *pszBuf;
-    int cchBufSize;
+    int32_t cchBufSize;
 };
 extern void cppISteamNetworkingUtils_SteamNetworkingUtils001_ConvertPingLocationToString( struct cppISteamNetworkingUtils_SteamNetworkingUtils001_ConvertPingLocationToString_params *params );
 
@@ -59,7 +59,7 @@ extern void cppISteamNetworkingUtils_SteamNetworkingUtils001_IsPingMeasurementIn
 struct cppISteamNetworkingUtils_SteamNetworkingUtils001_GetPingToDataCenter_params
 {
     void *linux_side;
-    int _ret;
+    int32_t _ret;
     SteamNetworkingPOPID popID;
     SteamNetworkingPOPID *pViaRelayPoP;
 };
@@ -68,7 +68,7 @@ extern void cppISteamNetworkingUtils_SteamNetworkingUtils001_GetPingToDataCenter
 struct cppISteamNetworkingUtils_SteamNetworkingUtils001_GetDirectPingToPOP_params
 {
     void *linux_side;
-    int _ret;
+    int32_t _ret;
     SteamNetworkingPOPID popID;
 };
 extern void cppISteamNetworkingUtils_SteamNetworkingUtils001_GetDirectPingToPOP( struct cppISteamNetworkingUtils_SteamNetworkingUtils001_GetDirectPingToPOP_params *params );
@@ -76,16 +76,16 @@ extern void cppISteamNetworkingUtils_SteamNetworkingUtils001_GetDirectPingToPOP(
 struct cppISteamNetworkingUtils_SteamNetworkingUtils001_GetPOPCount_params
 {
     void *linux_side;
-    int _ret;
+    int32_t _ret;
 };
 extern void cppISteamNetworkingUtils_SteamNetworkingUtils001_GetPOPCount( struct cppISteamNetworkingUtils_SteamNetworkingUtils001_GetPOPCount_params *params );
 
 struct cppISteamNetworkingUtils_SteamNetworkingUtils001_GetPOPList_params
 {
     void *linux_side;
-    int _ret;
+    int32_t _ret;
     SteamNetworkingPOPID *list;
-    int nListSz;
+    int32_t nListSz;
 };
 extern void cppISteamNetworkingUtils_SteamNetworkingUtils001_GetPOPList( struct cppISteamNetworkingUtils_SteamNetworkingUtils001_GetPOPList_params *params );
 
@@ -110,7 +110,7 @@ struct cppISteamNetworkingUtils_SteamNetworkingUtils001_SetConfigValue_params
     bool _ret;
     uint32_t eValue;
     uint32_t eScopeType;
-    intptr_t scopeObj;
+    int32_t scopeObj;
     uint32_t eDataType;
     const void *pArg;
 };
@@ -122,7 +122,7 @@ struct cppISteamNetworkingUtils_SteamNetworkingUtils001_GetConfigValue_params
     uint32_t _ret;
     uint32_t eValue;
     uint32_t eScopeType;
-    intptr_t scopeObj;
+    int32_t scopeObj;
     uint32_t *pOutDataType;
     void *pResult;
     size_t *cbResult;
